Remove mostly useless OpenSSL compatability check.
Based on Debian patch.

Don't check status field of the OpenSSL version

There is no reason to check the version of
OpenSSL on essentially any modern UNIX system.

If it's not compatible, the soname will change.
OpenSSH seems to want to do a check for the
soname based on the version number but then wants
to keep the status of the release the same.

Remove that check on the status since it doesn't
tell you anything usefull about how actually
compatible that version is.

Author: Colin Watson <cjwatson@debian.org>
Bug-Debian: https://bugs.debian.org/93581
Bug-Debian: https://bugs.debian.org/664383
Bug-Debian: https://bugs.debian.org/732940