Add/restore TCP wrappers support
Based on Debian patch.

They note: Support for TCP wrappers was dropped
in OpenSSH 6.7. See this message and thread:

https://lists.mindrot.org/pipermail/openssh-unix-dev/2014-April/032497.html

It is true that this reduces preauth attack
surface in sshd. On the other hand, this support
seems to be widely used, and abruptly dropping it
(from the perspective of users who don't read the
openssh-unix-dev mailing lists) could easily
cause problems in practice. While it is not
entirely clear what the right long-term answer
is, it likely doesn't involve abruptly dropping
features.

I will also note the TCP wrappers is a very
mature yet well maintained software package,
has been reviewed and audited extensively, and
is unlikely (in my opinion only, of course) to
introduce more security issues than it solves.