"LogLevel SILENT" compatibility
Based on Debian patch.

"LogLevel SILENT" (-qq) was introduced in Debian
openssh 1:3.0.1p1-1 to match the behaviour of the
commercial/non-free SSH implementation, in which
"-q" does not suppress fatal errors. This was
unintentionally broken in 1:4.6p1-2. The parts
that remain are basic configuration file
compatibility, and an adjustment to
"Pseudo-terminal will not be allocated ...".

Low impact patch, enhances drop-in compatibility
for Debian systems.

Author: Matthew Vernon <matthew@debian.org>
Author: Colin Watson <cjwatson@debian.org>