Backport 41935882f4e82de60dbd6e033eabe79e1b963518
upstream: When doing an sftp recursive upload or
download of a read-only directory, ensure that
the directory is created with write and execute
permissions in the interim so that we can
actually complete the transfer, then set the
directory permission as the final step.

(The execute bit is only likely to be an issue
with a non-POSIX server). [bz#3222, ok djm]

OpenBSD-Commit-ID: a82606212f2796e31f0e1af94a63355a7ad5d903