Backport and merge upstream pending PR #231
https://github.com/openssh/openssh-portable/pull/231
https://github.com/openssh/openssh-portable/pull/231/commits/6c6813f608f4d6046524e3ea4899e9fc53421b36
https://github.com/openssh/openssh-portable/pull/231/commits/de28eef2943c95b25e5bdc837f56d9859bb424b7
https://github.com/openssh/openssh-portable/pull/231/commits/aac32c917f8d74a546a5ca731e9bcdf8e9987e21
https://github.com/openssh/openssh-portable/pull/231/commits/0f7dd4d1d08cb097165f7ecc80a79eb4c53b606e
https://marc.info/?l=openssh-unix-dev&m=147248154926769
https://marc.info/?l=openssh-unix-dev&m=161301825023655
https://marc.info/?l=openssh-unix-dev&m=161387765919633

Description (from @vog):

Dear OpenSSH developers,

I kindly ask you to review the attached set of
patches which introduce ssh_config equivalents
of the flags -N, -n and -f in a
straight-forward way:

    NoShell for -N
    StdinNull for -n
    ForkAfterAuthentication for -f

The ssh_config names were derived directly from
the internal flag names, e.g. no_shell_flag was
moved to option.no_shell and hence the
ssh_config name became NoShell.

The man pages ssh(1) and ssh_config(5) are
adjusted accordingly.

As a final remark, I noticed that the variable
ono_shell_flag (not to be confused with
no_shell_flag) is only assigned once and never
used. So I assume this is dead code and I'm
proposing a patch to remove it.