Backport d98f14b5328922ae3085e07007d820c4f655b57a
Upstream patch.

UpdateHostkeys: better CheckHostIP handling

When preparing to update the known_hosts file,
fully check both entries for both the host and
the address (if CheckHostIP enabled) and ensure
that, at the end of the operation, entries for
both are recorded.

Make sure this works with HashKnownHosts too,
which requires maintaining a list of entry-types
seen across the whole file for each key.

OK: markus@

OpenBSD-Commit-ID: 374dc263103f6b343d9671f87dbf81ffd0d6abdd