Add a fix for upstream bug bz#2556
Currently sshd exits with fatal error if it see
ChrootDirectory option when running as non-root.

This is wrong on Linux as there non-root process
can perform chroot as long as it has SYS_CHROOT
effective capability. So the code should either
query the capability or the check should be
removed as sshd treats any chroot syscall errors
as fatal in any case.

UID 0 check removed.